Move window class to its own file
