+++ Editor/TextMeshProPromptWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92a310ef501e5d349b6151177f48c5e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
